Fixed warning triggered by incorrect field data type

+++ dbmodels.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 from sqlmodel import Field, SQLModel, Column, Integer, DECIMAL
+from decimal import Decimal
 
 
 class Product(SQLModel, table=True, arbitrary_types_allowed=True):
@@ -9,6 +10,6 @@ class Product(SQLModel, table=True, arbitrary_types_allowed=True):
     brand: str
     image_url: str
     description: str
-    price: int = Field(sa_column=Column(DECIMAL(100, 2)))
+    price: Decimal = Field(sa_column=Column(DECIMAL(100, 2)))
     supermarket: str
     last_updated: datetime
